screen for taking the route

+++ app/src/main/java/ru/nn/tripnn/data/dto/DistanceDto.kt
@@ -1,7 +1,7 @@
 package ru.nn.tripnn.data.dto
 
 data class DistanceDto (
-    val distance: Double,
+    val distance: Int,
     val duration: Int,
     val isAvailable: Boolean
 )
